FEAT/CLEANUP: Several identifiers, class names, test utilities, comments, and constants were ported verbatim from device-detection-java and still reference device-detection-specific concepts (Hash algorithm, User-Agent, Client Hints, TAC) that have no relevance to IP Intelligence. Change where appropriate and deprecate items that aren't relevenat to ipi and aren't core to the solution.
