+++ ip-intelligence.cloud/src/main/java/fiftyone/ipintelligence/cloud/flowelements/HardwareProfileCloudEngine.java
@@ -35,7 +35,11 @@ import org.slf4j.Logger;
  * Engine that takes the JSON response from the {@link CloudRequestEngine} and
  * uses it populate a {@link MultiIPIDataCloud} instance for easier consumption.
  * @see <a href="https://github.com/51Degrees/specifications/blob/main/ip-intelligence-specification/pipeline-elements/hardware-profile-lookup-cloud.md">Specification</a>
+ * @deprecated hardware profile lookup is a device detection feature and has
+ * no equivalent in IP Intelligence. This engine will be removed in a future
+ * version.
  */
+@Deprecated
 public class HardwareProfileCloudEngine
     extends PropertyKeyedCloudEngineBase<MultiIPIDataCloud, IPIntelligenceData> {
