+++ ip-intelligence.cloud/src/test/java/fiftyone/ipintelligence/cloud/TestsBase.java
@@ -22,7 +22,7 @@
 
 package fiftyone.ipintelligence.cloud;
 
-import fiftyone.ipintelligence.shared.testhelpers.UserAgentGenerator;
+import fiftyone.ipintelligence.shared.testhelpers.IpAddressGenerator;
 import fiftyone.pipeline.util.FileFinder;
 
 import static fiftyone.ipintelligence.shared.testhelpers.FileUtils.IP_ADDRESSES_FILE_NAME;
@@ -30,19 +30,19 @@ import static fiftyone.ipintelligence.shared.testhelpers.FileUtils.IP_ADDRESSES_
 public class TestsBase {
 
     private WrapperCloud wrapper = null;
-    private UserAgentGenerator userAgents;
+    private IpAddressGenerator ipAddresses;
 
     protected WrapperCloud getWrapper() {
         return wrapper;
     }
 
-    protected UserAgentGenerator getUserAgents() {
-        return userAgents;
+    protected IpAddressGenerator getIpAddresses() {
+        return ipAddresses;
     }
 
     protected void testInitialize() throws Exception {
         wrapper = new WrapperCloud();
-        userAgents = new UserAgentGenerator(
+        ipAddresses = new IpAddressGenerator(
             FileFinder.getFilePath(IP_ADDRESSES_FILE_NAME));
     }
