+++ ip-intelligence.cloud/src/test/java/fiftyone/ipintelligence/cloud/data/ValueCloudTests.java
@@ -24,7 +24,6 @@ package fiftyone.ipintelligence.cloud.data;
 
 import fiftyone.ipintelligence.cloud.TestsBase;
 import fiftyone.ipintelligence.cloud.ValueTests;
-import fiftyone.ipintelligence.shared.testhelpers.UserAgentGenerator;
 
 import org.junit.After;
 import org.junit.Before;
@@ -34,8 +33,6 @@ import org.junit.Test;
 @Ignore
 public class ValueCloudTests extends TestsBase {
 
-    	private UserAgentGenerator userAgents;
-    	
         @Before
         public void init() throws Exception {
             testInitialize();
