+++ ip-intelligence.engine.on-premise/src/main/java/fiftyone/ipintelligence/engine/onpremise/data/IPIntelligenceDataHash.java
@@ -23,23 +23,14 @@
 package fiftyone.ipintelligence.engine.onpremise.data;
 
 import fiftyone.ipintelligence.shared.IPIntelligenceData;
-import fiftyone.pipeline.engines.data.AspectPropertyValue;
-
-import java.util.List;
 
 /**
  * On-premise specific interface for {@link IPIntelligenceData}.
- * @see <a href="https://github.com/51Degrees/specifications/blob/main/ip-intelligence-specification/pipeline-elements/ip-intelligence-on-premise.md#element-data">Specification</a>
+ * @deprecated "Hash" refers to the device detection algorithm and has no
+ * meaning for IP Intelligence. Use {@link IPIntelligenceDataOnPremise}
+ * instead; this interface only remains for backward compatibility and will
+ * be removed in a future version.
  */
-public interface IPIntelligenceDataHash extends IPIntelligenceData {
-
-    /**
-     * Get the values for the specified property as a List&lt;&gt;.
-     * @param propertyName name of the property to get values for
-     * @param parametrizedTypes types parametrizing {@link AspectPropertyValue} starting with {@link List}
-     * @return values as a list
-     */
-    AspectPropertyValue<List<?>> getValues(
-            String propertyName,
-            Class<?>[] parametrizedTypes);
+@Deprecated
+public interface IPIntelligenceDataHash extends IPIntelligenceDataOnPremise {
 }
