+++ ip-intelligence.engine.on-premise/src/test/java/fiftyone/ipintelligence/engine/onpremise/TestsBase.java
@@ -22,7 +22,7 @@
 
 package fiftyone.ipintelligence.engine.onpremise;
 
-import fiftyone.ipintelligence.shared.testhelpers.UserAgentGenerator;
+import fiftyone.ipintelligence.shared.testhelpers.IpAddressGenerator;
 import fiftyone.ipintelligence.shared.testhelpers.FileUtils;
 import fiftyone.pipeline.engines.Constants;
 
@@ -31,22 +31,22 @@ import static fiftyone.pipeline.util.FileFinder.getFilePath;
 
 public class TestsBase {
 
-    private WrapperHash wrapper = null;
-    private UserAgentGenerator userAgents;
+    private WrapperOnPremise wrapper = null;
+    private IpAddressGenerator ipAddresses;
 
-    protected WrapperHash getWrapper() {
+    protected WrapperOnPremise getWrapper() {
         return wrapper;
     }
 
-    protected UserAgentGenerator getUserAgents() {
-        return userAgents;
+    protected IpAddressGenerator getIpAddresses() {
+        return ipAddresses;
     }
 
     protected void testInitialize(Constants.PerformanceProfiles profile) throws Exception {
-        wrapper = new WrapperHash(
-            FileUtils.getHashFile(),
+        wrapper = new WrapperOnPremise(
+            FileUtils.getIpiFile(),
             profile);
-        userAgents = new UserAgentGenerator(
+        ipAddresses = new IpAddressGenerator(
             getFilePath(IP_ADDRESSES_FILE_NAME));
     }
