+++ ip-intelligence.shared/src/test/java/fiftyone/ipintelligence/shared/testhelpers/flowelements/ProcessTests.java
@@ -43,7 +43,7 @@ public class ProcessTests {
         }
     }
 
-    public static void emptyUserAgent(
+    public static void emptyIpAddress(
         Wrapper wrapper,
         DataValidator validator) throws Exception {
         try (FlowData data = wrapper.getPipeline().createFlowData()) {
