+++ ip-intelligence/src/test/java/fiftyone/ipintelligence/CloudRequestOriginTests.java
@@ -94,11 +94,7 @@ public class CloudRequestOriginTests {
                 .build();
 
             try(FlowData data = pipeline.createFlowData()) {        
-                data.addEvidence("query.client-ip",                         
-                        "Mozilla/5.0 (Linux; Android 8.0.0; Pixel 2 XL " +
-                        "Build/OPD1.170816.004) AppleWebKit/537.36 " +
-                        "(KHTML, like Gecko) Chrome/80.0.3987.106 " +
-                        "Mobile Safari/537.36")
+                data.addEvidence("query.client-ip", "50.154.29.201")
                     .process();
             }
         }
